[ObjC] Adapt the varint size calculation from C++
The C++ has a version that is branchless and also very short, so just use that.

However, don't bother with also doing the Intel version as that won't likely
matter for Apple platforms for much longer.

We've seen cases where folks use clang flags to aggressively inline things and
so this could actual shrink things depending on how many times it gets inlined
within CodedOutputStream.

PiperOrigin-RevId: 830517346